predict the classes using trained algorithms and give cnfusion matrix in return

predict the classes using trained algorithms and give cnfusion matrix in return

BuildPrediction(x)

Arguments

x
input is a dataframe for bag of word file in which columns are the terms and row are binary variable 1 if that term exist in that data instance

Value

3 confusion matrix for each trained classification algorithm

Examples

library(SentiAnalyzer)
csv_data <- read.csv(system.file(package = "SentiAnalyzer", "extdata/testing.csv"))
my_training_data <- BuildPrediction(csv_data)

#> Warning: package 'caret' was built under R version 3.5.3

#> Loading required package: lattice

#> Loading required package: ggplot2

#> Warning: package 'ggplot2' was built under R version 3.5.2

#> Warning: model fit failed for Fold01: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold01: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold02: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold02: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold03: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold03: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold04: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold04: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold05: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt, worst

#> Warning: model fit failed for Fold05: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt, worst

#> Warning: model fit failed for Fold06: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: atmospher, awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold06: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: atmospher, awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold07: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold07: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold08: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold08: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold09: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold09: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, minut, perfect, wasnt

#> Warning: model fit failed for Fold10: fL=1, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, manag, minut, perfect, wasnt

#> Warning: model fit failed for Fold10: fL=2, usekernel=FALSE, adjust=0.5 Error in NaiveBayes.default(x, y, usekernel = FALSE, fL = param$fL, ...) : 
#>   Zero variances for at least one class in variables: awesom, bad, bland, delici, fantast, great, manag, minut, perfect, wasnt

#> Warning: There were missing values in resampled performance measures.

#> Warning: missing values found in aggregated results